feat: refactorización hook
